+++ pygeoapi/api/__init__.py
@@ -34,7 +34,9 @@
 # OTHER DEALINGS IN THE SOFTWARE.
 #
 # =================================================================
-""" Root level code of pygeoapi, parsing content provided by web framework.
+
+"""
+Root level code of pygeoapi, parsing content provided by web framework.
 Returns content from plugins and sets responses.
 """
 
@@ -48,34 +50,23 @@ from http import HTTPStatus
 import logging
 import re
 from typing import Any, Tuple, Union, Optional
-import urllib.parse
 
 from dateutil.parser import parse as dateparse
-from pygeofilter.parsers.ecql import parse as parse_ecql_text
-from pygeofilter.parsers.cql_json import parse as parse_cql_json
-from pyproj.exceptions import CRSError
 import pytz
 
 from pygeoapi import __version__, l10n
-from pygeoapi.formatter.base import FormatterSerializationError
-from pygeoapi.linked_data import (geojson2jsonld, jsonldify,
-                                  jsonldify_collection)
+from pygeoapi.linked_data import jsonldify, jsonldify_collection
 from pygeoapi.log import setup_logger
+from pygeoapi.plugin import load_plugin
 from pygeoapi.process.manager.base import get_manager
-from pygeoapi.plugin import load_plugin, PLUGINS
-from pygeoapi.provider.base import (
-    ProviderGenericError, ProviderConnectionError, ProviderTypeError)
+from pygeoapi.provider.base import ProviderConnectionError, ProviderTypeError
 
-from pygeoapi.models.cql import CQLModel
-from pygeoapi.util import (dategetter,
-                           UrlPrefetcher,
-                           filter_dict_by_key_value, get_provider_by_type,
-                           get_provider_default, get_typed_value,
-                           render_j2_template, str2bool,
-                           TEMPLATES, to_json, get_api_rules, get_base_url,
-                           get_crs_from_uri, get_supported_crs_list,
-                           modify_pygeofilter, CrsTransformSpec,
-                           transform_bbox)
+from pygeoapi.util import (
+    CrsTransformSpec, TEMPLATES, UrlPrefetcher, dategetter,
+    filter_dict_by_key_value, get_api_rules, get_base_url,
+    get_provider_by_type, get_provider_default, get_typed_value,
+    get_crs_from_uri, get_supported_crs_list, render_j2_template, to_json
+)
 
 LOGGER = logging.getLogger(__name__)
 
@@ -109,65 +100,14 @@ FORMAT_TYPES = OrderedDict((
 #: Locale used for system responses (e.g. exceptions)
 SYSTEM_LOCALE = l10n.Locale('en', 'US')
 
-CONFORMANCE = {
-    'common': [
-        'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/core',
-        'http://www.opengis.net/spec/ogcapi-common-2/1.0/conf/collections',
-        'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/landing-page',
-        'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/json',
-        'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/html',
-        'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/oas30'
-    ],
-    'feature': [
-        'http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core',
-        'http://www.opengis.net/spec/ogcapi-features-1/1.0/req/oas30',
-        'http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/html',
-        'http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson',
-        'http://www.opengis.net/spec/ogcapi-features-2/1.0/conf/crs',
-        'http://www.opengis.net/spec/ogcapi-features-3/1.0/conf/queryables',
-        'http://www.opengis.net/spec/ogcapi-features-3/1.0/conf/queryables-query-parameters',  # noqa
-        'http://www.opengis.net/spec/ogcapi-features-4/1.0/conf/create-replace-delete',  # noqa
-        'http://www.opengis.net/spec/ogcapi-features-5/1.0/conf/schemas',
-        'http://www.opengis.net/spec/ogcapi-features-5/1.0/req/core-roles-features' # noqa
-    ],
-    'coverage': [
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/core',
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/oas30',
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/html',
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/geodata-coverage',  # noqa
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/coverage-subset',  # noqa
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/coverage-rangesubset',  # noqa
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/coverage-bbox',  # noqa
-        'http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/coverage-datetime'  # noqa
-    ],
-    'map': [
-        'http://www.opengis.net/spec/ogcapi-maps-1/1.0/conf/core'
-    ],
-    'tile': [
-        'http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/core',
-        'http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/mvt',
-        'http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/tileset',
-        'http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/tilesets-list',
-        'http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/oas30',
-        'http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/geodata-tilesets'
-    ],
-    'record': [
-        'http://www.opengis.net/spec/ogcapi-records-1/1.0/conf/core',
-        'http://www.opengis.net/spec/ogcapi-records-1/1.0/conf/sorting',
-        'http://www.opengis.net/spec/ogcapi-records-1/1.0/conf/opensearch',
-        'http://www.opengis.net/spec/ogcapi-records-1/1.0/conf/json',
-        'http://www.opengis.net/spec/ogcapi-records-1/1.0/conf/html'
-    ],
-    'process': [
-        'http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/ogc-process-description', # noqa
-        'http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/core',
-        'http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/json',
-        'http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/oas30'
-    ],
-    'edr': [
-        'http://www.opengis.net/spec/ogcapi-edr-1/1.0/conf/core'
-    ]
-}
+CONFORMANCE_CLASSES = [
+    'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/core',
+    'http://www.opengis.net/spec/ogcapi-common-2/1.0/conf/collections',
+    'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/landing-page',
+    'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/json',
+    'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/html',
+    'http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/oas30'
+]
 
 OGC_RELTYPES_BASE = 'http://www.opengis.net/def/rel/ogc/1.0'
 
@@ -180,11 +120,27 @@ DEFAULT_CRS = 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'
 DEFAULT_STORAGE_CRS = DEFAULT_CRS
 
 
-def all_apis():
-    # NOTE: this is a function and not a constant to avoid import loops
-    from . import environmental_data_retrieval, maps, processes, stac
+def all_apis() -> dict:
+    """
+    Return all supported API modules
 
-    return [environmental_data_retrieval, maps, processes, stac]
+    NOTE: this is a function and not a constant to avoid import loops
+
+    :returns: `dict` of API provider type, API module
+    """
+
+    from . import (coverages, environmental_data_retrieval, itemtypes, maps,
+                   processes, tiles, stac)
+
+    return {
+        'coverage': coverages,
+        'edr': environmental_data_retrieval,
+        'itemtypes': itemtypes,
+        'map': maps,
+        'process': processes,
+        'tile': tiles,
+        'stac': stac
+    }
 
 
 def pre_process(func):
@@ -895,21 +851,31 @@ class API:
         :returns: tuple of headers, status code, content
         """
 
+        apis_dict = all_apis()
+
         if not request.is_valid():
             return self.get_format_exception(request)
 
-        conformance_list = CONFORMANCE['common']
+        conformance_list = CONFORMANCE_CLASSES
 
         for key, value in self.config['resources'].items():
             if value['type'] == 'process':
-                conformance_list.extend(CONFORMANCE[value['type']])
+                conformance_list.extend(
+                    apis_dict['process'].CONFORMANCE_CLASSES)
             else:
                 for provider in value['providers']:
-                    if provider['type'] in CONFORMANCE:
-                        conformance_list.extend(CONFORMANCE[provider['type']])
+                    if provider['type'] in apis_dict:
+                        conformance_list.extend(
+                            apis_dict[provider['type']].CONFORMANCE_CLASSES)
+                    if provider['type'] == 'feature':
+                        conformance_list.extend(
+                            apis_dict['itemtypes'].CONFORMANCE_CLASSES_FEATURES)  # noqa
+                    if provider['type'] == 'record':
+                        conformance_list.extend(
+                            apis_dict['itemtypes'].CONFORMANCE_CLASSES_RECORDS)
 
         conformance = {
-            'conformsTo': list(set(conformance_list))
+            'conformsTo': sorted(list(set(conformance_list)))
         }
 
         headers = request.get_response_headers(**self.api_headers)
@@ -1316,1401 +1282,6 @@ class API:
             return headers, HTTPStatus.OK, to_json(jsonld, self.pretty_print)
 
         return headers, HTTPStatus.OK, to_json(fcm, self.pretty_print)
-
-    @gzip
-    @pre_process
-    @jsonldify
-    def get_collection_schema(
-            self, request: Union[APIRequest, Any],
-            dataset) -> Tuple[dict, int, str]:
-        """
-        Returns a collection schema
-
-        :param request: A request object
-        :param dataset: dataset name
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid():
-            return self.get_format_exception(request)
-        headers = request.get_response_headers(**self.api_headers)
-
-        if any([dataset is None,
-                dataset not in self.config['resources'].keys()]):
-
-            msg = 'Collection not found'
-            return self.get_exception(
-                HTTPStatus.NOT_FOUND, headers, request.format, 'NotFound', msg)
-
-        LOGGER.debug('Creating collection schema')
-        try:
-            LOGGER.debug('Loading feature provider')
-            p = load_plugin('provider', get_provider_by_type(
-                self.config['resources'][dataset]['providers'], 'feature'))
-        except ProviderTypeError:
-            try:
-                LOGGER.debug('Loading coverage provider')
-                p = load_plugin('provider', get_provider_by_type(
-                    self.config['resources'][dataset]['providers'], 'coverage'))  # noqa
-            except ProviderTypeError:
-                LOGGER.debug('Loading record provider')
-                p = load_plugin('provider', get_provider_by_type(
-                    self.config['resources'][dataset]['providers'], 'record'))
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        schema = {
-            'type': 'object',
-            'title': l10n.translate(
-                self.config['resources'][dataset]['title'], request.locale),
-            'properties': {},
-            '$schema': 'http://json-schema.org/draft/2019-09/schema',
-            '$id': f'{self.get_collections_url()}/{dataset}/schema'
-        }
-
-        if p.type != 'coverage':
-            schema['properties']['geometry'] = {
-                '$ref': 'https://geojson.org/schema/Geometry.json',
-                'x-ogc-role': 'primary-geometry'
-            }
-
-        for k, v in p.fields.items():
-            schema['properties'][k] = v
-
-            if k == p.id_field:
-                schema['properties'][k]['x-ogc-role'] = 'id'
-            if k == p.time_field:
-                schema['properties'][k]['x-ogc-role'] = 'primary-instant'
-
-        if request.format == F_HTML:  # render
-            schema['title'] = l10n.translate(
-                self.config['resources'][dataset]['title'], request.locale)
-
-            schema['collections_path'] = self.get_collections_url()
-
-            content = render_j2_template(self.tpl_config,
-                                         'collections/schema.html',
-                                         schema, request.locale)
-
-            return headers, HTTPStatus.OK, content
-
-        headers['Content-Type'] = 'application/schema+json'
-
-        return headers, HTTPStatus.OK, to_json(schema, self.pretty_print)
-
-    @gzip
-    @pre_process
-    @jsonldify
-    def get_collection_queryables(self, request: Union[APIRequest, Any],
-                                  dataset=None) -> Tuple[dict, int, str]:
-        """
-        Provide collection queryables
-
-        :param request: A request object
-        :param dataset: name of collection
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid():
-            return self.get_format_exception(request)
-        headers = request.get_response_headers(**self.api_headers)
-
-        if any([dataset is None,
-                dataset not in self.config['resources'].keys()]):
-
-            msg = 'Collection not found'
-            return self.get_exception(
-                HTTPStatus.NOT_FOUND, headers, request.format, 'NotFound', msg)
-
-        LOGGER.debug('Creating collection queryables')
-        try:
-            LOGGER.debug('Loading feature provider')
-            p = load_plugin('provider', get_provider_by_type(
-                self.config['resources'][dataset]['providers'], 'feature'))
-        except ProviderTypeError:
-            try:
-                LOGGER.debug('Loading coverage provider')
-                p = load_plugin('provider', get_provider_by_type(
-                    self.config['resources'][dataset]['providers'], 'coverage'))  # noqa
-            except ProviderTypeError:
-                LOGGER.debug('Loading record provider')
-                p = load_plugin('provider', get_provider_by_type(
-                    self.config['resources'][dataset]['providers'], 'record'))
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        queryables = {
-            'type': 'object',
-            'title': l10n.translate(
-                self.config['resources'][dataset]['title'], request.locale),
-            'properties': {},
-            '$schema': 'http://json-schema.org/draft/2019-09/schema',
-            '$id': f'{self.get_collections_url()}/{dataset}/queryables'
-        }
-
-        if p.fields:
-            queryables['properties']['geometry'] = {
-                '$ref': 'https://geojson.org/schema/Geometry.json',
-                'x-ogc-role': 'primary-geometry'
-            }
-
-        for k, v in p.fields.items():
-            show_field = False
-            if p.properties:
-                if k in p.properties:
-                    show_field = True
-            else:
-                show_field = True
-
-            if show_field:
-                queryables['properties'][k] = {
-                    'title': k,
-                    'type': v['type']
-                }
-                if 'values' in v:
-                    queryables['properties'][k]['enum'] = v['values']
-
-                if k == p.id_field:
-                    queryables['properties'][k]['x-ogc-role'] = 'id'
-                if k == p.time_field:
-                    queryables['properties'][k]['x-ogc-role'] = 'primary-instant'  # noqa
-
-        if request.format == F_HTML:  # render
-            queryables['title'] = l10n.translate(
-                self.config['resources'][dataset]['title'], request.locale)
-
-            queryables['collections_path'] = self.get_collections_url()
-
-            content = render_j2_template(self.tpl_config,
-                                         'collections/queryables.html',
-                                         queryables, request.locale)
-
-            return headers, HTTPStatus.OK, content
-
-        headers['Content-Type'] = 'application/schema+json'
-
-        return headers, HTTPStatus.OK, to_json(queryables, self.pretty_print)
-
-    @gzip
-    @pre_process
-    def get_collection_items(
-            self, request: Union[APIRequest, Any],
-            dataset) -> Tuple[dict, int, str]:
-        """
-        Queries collection
-
-        :param request: A request object
-        :param dataset: dataset name
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid(PLUGINS['formatter'].keys()):
-            return self.get_format_exception(request)
-
-        # Set Content-Language to system locale until provider locale
-        # has been determined
-        headers = request.get_response_headers(SYSTEM_LOCALE,
-                                               **self.api_headers)
-
-        properties = []
-        reserved_fieldnames = ['bbox', 'bbox-crs', 'crs', 'f', 'lang', 'limit',
-                               'offset', 'resulttype', 'datetime', 'sortby',
-                               'properties', 'skipGeometry', 'q',
-                               'filter', 'filter-lang', 'filter-crs']
-
-        collections = filter_dict_by_key_value(self.config['resources'],
-                                               'type', 'collection')
-
-        if dataset not in collections.keys():
-            msg = 'Collection not found'
-            return self.get_exception(
-                HTTPStatus.NOT_FOUND, headers, request.format, 'NotFound', msg)
-
-        LOGGER.debug('Processing query parameters')
-
-        LOGGER.debug('Processing offset parameter')
-        try:
-            offset = int(request.params.get('offset'))
-            if offset < 0:
-                msg = 'offset value should be positive or zero'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        except TypeError as err:
-            LOGGER.warning(err)
-            offset = 0
-        except ValueError:
-            msg = 'offset value should be an integer'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        LOGGER.debug('Processing limit parameter')
-        try:
-            limit = int(request.params.get('limit'))
-            # TODO: We should do more validation, against the min and max
-            #       allowed by the server configuration
-            if limit <= 0:
-                msg = 'limit value should be strictly positive'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        except TypeError as err:
-            LOGGER.warning(err)
-            limit = int(self.config['server']['limit'])
-        except ValueError:
-            msg = 'limit value should be an integer'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        resulttype = request.params.get('resulttype') or 'results'
-
-        LOGGER.debug('Processing bbox parameter')
-
-        bbox = request.params.get('bbox')
-
-        if bbox is None:
-            bbox = []
-        else:
-            try:
-                bbox = validate_bbox(bbox)
-            except ValueError as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-
-        LOGGER.debug('Processing datetime parameter')
-        datetime_ = request.params.get('datetime')
-        try:
-            datetime_ = validate_datetime(collections[dataset]['extents'],
-                                          datetime_)
-        except ValueError as err:
-            msg = str(err)
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        LOGGER.debug('processing q parameter')
-        q = request.params.get('q') or None
-
-        LOGGER.debug('Loading provider')
-
-        provider_def = None
-        try:
-            provider_type = 'feature'
-            provider_def = get_provider_by_type(
-                collections[dataset]['providers'], provider_type)
-            p = load_plugin('provider', provider_def)
-        except ProviderTypeError:
-            try:
-                provider_type = 'record'
-                provider_def = get_provider_by_type(
-                    collections[dataset]['providers'], provider_type)
-                p = load_plugin('provider', provider_def)
-            except ProviderTypeError:
-                msg = 'Invalid provider type'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'NoApplicableCode', msg)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        crs_transform_spec = None
-        if provider_type == 'feature':
-            # crs query parameter is only available for OGC API - Features
-            # right now, not for OGC API - Records.
-            LOGGER.debug('Processing crs parameter')
-            query_crs_uri = request.params.get('crs')
-            try:
-                crs_transform_spec = self._create_crs_transform_spec(
-                    provider_def, query_crs_uri,
-                )
-            except (ValueError, CRSError) as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-            self._set_content_crs_header(headers, provider_def, query_crs_uri)
-
-        LOGGER.debug('Processing bbox-crs parameter')
-        bbox_crs = request.params.get('bbox-crs')
-        if bbox_crs is not None:
-            # Validate bbox-crs parameter
-            if len(bbox) == 0:
-                msg = 'bbox-crs specified without bbox parameter'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'NoApplicableCode', msg)
-
-            if len(bbox_crs) == 0:
-                msg = 'bbox-crs specified but is empty'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'NoApplicableCode', msg)
-
-            supported_crs_list = get_supported_crs_list(provider_def, DEFAULT_CRS_LIST) # noqa
-            if bbox_crs not in supported_crs_list:
-                msg = f'bbox-crs {bbox_crs} not supported for this collection'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'NoApplicableCode', msg)
-        elif len(bbox) > 0:
-            # bbox but no bbox-crs parm: assume bbox is in default CRS
-            bbox_crs = DEFAULT_CRS
-
-        # Transform bbox to storageCRS
-        # when bbox-crs different from storageCRS.
-        if len(bbox) > 0:
-            try:
-                # Get a pyproj CRS instance for the Collection's Storage CRS
-                storage_crs = provider_def.get('storage_crs', DEFAULT_STORAGE_CRS) # noqa
-
-                # Do the (optional) Transform to the Storage CRS
-                bbox = transform_bbox(bbox, bbox_crs, storage_crs)
-            except CRSError as e:
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'NoApplicableCode', str(e))
-
-        LOGGER.debug('processing property parameters')
-        for k, v in request.params.items():
-            if k not in reserved_fieldnames and k in list(p.fields.keys()):
-                LOGGER.debug(f'Adding property filter {k}={v}')
-                properties.append((k, v))
-
-        LOGGER.debug('processing sort parameter')
-        val = request.params.get('sortby')
-
-        if val is not None:
-            sortby = []
-            sorts = val.split(',')
-            for s in sorts:
-                prop = s
-                order = '+'
-                if s[0] in ['+', '-']:
-                    order = s[0]
-                    prop = s[1:]
-
-                if prop not in p.fields.keys():
-                    msg = 'bad sort property'
-                    return self.get_exception(
-                        HTTPStatus.BAD_REQUEST, headers, request.format,
-                        'InvalidParameterValue', msg)
-
-                sortby.append({'property': prop, 'order': order})
-        else:
-            sortby = []
-
-        LOGGER.debug('processing properties parameter')
-        val = request.params.get('properties')
-
-        if val is not None:
-            select_properties = val.split(',')
-            properties_to_check = set(p.properties) | set(p.fields.keys())
-
-            if (len(list(set(select_properties) -
-                         set(properties_to_check))) > 0):
-                msg = 'unknown properties specified'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        else:
-            select_properties = []
-
-        LOGGER.debug('processing skipGeometry parameter')
-        val = request.params.get('skipGeometry')
-        if val is not None:
-            skip_geometry = str2bool(val)
-        else:
-            skip_geometry = False
-
-        LOGGER.debug('Processing filter-crs parameter')
-        filter_crs_uri = request.params.get('filter-crs', DEFAULT_CRS)
-        LOGGER.debug('processing filter parameter')
-        cql_text = request.params.get('filter')
-        if cql_text is not None:
-            try:
-                filter_ = parse_ecql_text(cql_text)
-                filter_ = modify_pygeofilter(
-                    filter_,
-                    filter_crs_uri=filter_crs_uri,
-                    storage_crs_uri=provider_def.get('storage_crs'),
-                    geometry_column_name=provider_def.get('geom_field'),
-                )
-            except Exception as err:
-                LOGGER.error(err)
-                msg = f'Bad CQL string : {cql_text}'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        else:
-            filter_ = None
-
-        LOGGER.debug('Processing filter-lang parameter')
-        filter_lang = request.params.get('filter-lang')
-        # Currently only cql-text is handled, but it is optional
-        if filter_lang not in [None, 'cql-text']:
-            msg = 'Invalid filter language'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-        # Get provider locale (if any)
-        prv_locale = l10n.get_plugin_locale(provider_def, request.raw_locale)
-
-        LOGGER.debug('Querying provider')
-        LOGGER.debug(f'offset: {offset}')
-        LOGGER.debug(f'limit: {limit}')
-        LOGGER.debug(f'resulttype: {resulttype}')
-        LOGGER.debug(f'sortby: {sortby}')
-        LOGGER.debug(f'bbox: {bbox}')
-        if provider_type == 'feature':
-            LOGGER.debug(f'crs: {query_crs_uri}')
-        LOGGER.debug(f'datetime: {datetime_}')
-        LOGGER.debug(f'properties: {properties}')
-        LOGGER.debug(f'select properties: {select_properties}')
-        LOGGER.debug(f'skipGeometry: {skip_geometry}')
-        LOGGER.debug(f'language: {prv_locale}')
-        LOGGER.debug(f'q: {q}')
-        LOGGER.debug(f'cql_text: {cql_text}')
-        LOGGER.debug(f'filter_: {filter_}')
-        LOGGER.debug(f'filter-lang: {filter_lang}')
-        LOGGER.debug(f'filter-crs: {filter_crs_uri}')
-
-        try:
-            content = p.query(offset=offset, limit=limit,
-                              resulttype=resulttype, bbox=bbox,
-                              datetime_=datetime_, properties=properties,
-                              sortby=sortby, skip_geometry=skip_geometry,
-                              select_properties=select_properties,
-                              crs_transform_spec=crs_transform_spec,
-                              q=q, language=prv_locale, filterq=filter_)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        serialized_query_params = ''
-        for k, v in request.params.items():
-            if k not in ('f', 'offset'):
-                serialized_query_params += '&'
-                serialized_query_params += urllib.parse.quote(k, safe='')
-                serialized_query_params += '='
-                serialized_query_params += urllib.parse.quote(str(v), safe=',')
-
-        # TODO: translate titles
-        uri = f'{self.get_collections_url()}/{dataset}/items'
-        content['links'] = [{
-            'type': 'application/geo+json',
-            'rel': request.get_linkrel(F_JSON),
-            'title': 'This document as GeoJSON',
-            'href': f'{uri}?f={F_JSON}{serialized_query_params}'
-        }, {
-            'rel': request.get_linkrel(F_JSONLD),
-            'type': FORMAT_TYPES[F_JSONLD],
-            'title': 'This document as RDF (JSON-LD)',
-            'href': f'{uri}?f={F_JSONLD}{serialized_query_params}'
-        }, {
-            'type': FORMAT_TYPES[F_HTML],
-            'rel': request.get_linkrel(F_HTML),
-            'title': 'This document as HTML',
-            'href': f'{uri}?f={F_HTML}{serialized_query_params}'
-        }]
-
-        if offset > 0:
-            prev = max(0, offset - limit)
-            content['links'].append(
-                {
-                    'type': 'application/geo+json',
-                    'rel': 'prev',
-                    'title': 'items (prev)',
-                    'href': f'{uri}?offset={prev}{serialized_query_params}'
-                })
-
-        if 'numberMatched' in content:
-            if content['numberMatched'] > (limit + offset):
-                next_ = offset + limit
-                next_href = f'{uri}?offset={next_}{serialized_query_params}'
-                content['links'].append(
-                    {
-                        'type': 'application/geo+json',
-                        'rel': 'next',
-                        'title': 'items (next)',
-                        'href': next_href
-                    })
-
-        content['links'].append(
-            {
-                'type': FORMAT_TYPES[F_JSON],
-                'title': l10n.translate(
-                    collections[dataset]['title'], request.locale),
-                'rel': 'collection',
-                'href': uri
-            })
-
-        content['timeStamp'] = datetime.utcnow().strftime(
-            '%Y-%m-%dT%H:%M:%S.%fZ')
-
-        # Set response language to requested provider locale
-        # (if it supports language) and/or otherwise the requested pygeoapi
-        # locale (or fallback default locale)
-        l10n.set_response_language(headers, prv_locale, request.locale)
-
-        if request.format == F_HTML:  # render
-            # For constructing proper URIs to items
-
-            content['items_path'] = uri
-            content['dataset_path'] = '/'.join(uri.split('/')[:-1])
-            content['collections_path'] = self.get_collections_url()
-
-            content['offset'] = offset
-
-            content['id_field'] = p.id_field
-            if p.uri_field is not None:
-                content['uri_field'] = p.uri_field
-            if p.title_field is not None:
-                content['title_field'] = l10n.translate(p.title_field,
-                                                        request.locale)
-                # If title exists, use it as id in html templates
-                content['id_field'] = content['title_field']
-            content = render_j2_template(self.tpl_config,
-                                         'collections/items/index.html',
-                                         content, request.locale)
-            return headers, HTTPStatus.OK, content
-        elif request.format == 'csv':  # render
-            formatter = load_plugin('formatter',
-                                    {'name': 'CSV', 'geom': True})
-
-            try:
-                content = formatter.write(
-                    data=content,
-                    options={
-                        'provider_def': get_provider_by_type(
-                                            collections[dataset]['providers'],
-                                            'feature')
-                    }
-                )
-            except FormatterSerializationError as err:
-                LOGGER.error(err)
-                msg = 'Error serializing output'
-                return self.get_exception(
-                    HTTPStatus.INTERNAL_SERVER_ERROR, headers, request.format,
-                    'NoApplicableCode', msg)
-
-            headers['Content-Type'] = formatter.mimetype
-
-            if p.filename is None:
-                filename = f'{dataset}.csv'
-            else:
-                filename = f'{p.filename}'
-
-            cd = f'attachment; filename="{filename}"'
-            headers['Content-Disposition'] = cd
-
-            return headers, HTTPStatus.OK, content
-
-        elif request.format == F_JSONLD:
-            content = geojson2jsonld(
-                self, content, dataset, id_field=(p.uri_field or 'id')
-            )
-
-        return headers, HTTPStatus.OK, to_json(content, self.pretty_print)
-
-    @gzip
-    @pre_process
-    def post_collection_items(
-            self, request: Union[APIRequest, Any],
-            dataset) -> Tuple[dict, int, str]:
-        """
-        Queries collection or filter an item
-
-        :param request: A request object
-        :param dataset: dataset name
-
-        :returns: tuple of headers, status code, content
-        """
-
-        request_headers = request.headers
-
-        if not request.is_valid(PLUGINS['formatter'].keys()):
-            return self.get_format_exception(request)
-
-        # Set Content-Language to system locale until provider locale
-        # has been determined
-        headers = request.get_response_headers(SYSTEM_LOCALE,
-                                               **self.api_headers)
-
-        properties = []
-        reserved_fieldnames = ['bbox', 'f', 'limit', 'offset',
-                               'resulttype', 'datetime', 'sortby',
-                               'properties', 'skipGeometry', 'q',
-                               'filter-lang', 'filter-crs']
-
-        collections = filter_dict_by_key_value(self.config['resources'],
-                                               'type', 'collection')
-
-        if dataset not in collections.keys():
-            msg = 'Invalid collection'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        LOGGER.debug('Processing query parameters')
-
-        LOGGER.debug('Processing offset parameter')
-        try:
-            offset = int(request.params.get('offset'))
-            if offset < 0:
-                msg = 'offset value should be positive or zero'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        except TypeError as err:
-            LOGGER.warning(err)
-            offset = 0
-        except ValueError:
-            msg = 'offset value should be an integer'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        LOGGER.debug('Processing limit parameter')
-        try:
-            limit = int(request.params.get('limit'))
-            # TODO: We should do more validation, against the min and max
-            # allowed by the server configuration
-            if limit <= 0:
-                msg = 'limit value should be strictly positive'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        except TypeError as err:
-            LOGGER.warning(err)
-            limit = int(self.config['server']['limit'])
-        except ValueError:
-            msg = 'limit value should be an integer'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        resulttype = request.params.get('resulttype') or 'results'
-
-        LOGGER.debug('Processing bbox parameter')
-
-        bbox = request.params.get('bbox')
-
-        if bbox is None:
-            bbox = []
-        else:
-            try:
-                bbox = validate_bbox(bbox)
-            except ValueError as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-
-        LOGGER.debug('Processing datetime parameter')
-        datetime_ = request.params.get('datetime')
-        try:
-            datetime_ = validate_datetime(collections[dataset]['extents'],
-                                          datetime_)
-        except ValueError as err:
-            msg = str(err)
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        LOGGER.debug('processing q parameter')
-        val = request.params.get('q')
-
-        q = None
-        if val is not None:
-            q = val
-
-        LOGGER.debug('Loading provider')
-
-        try:
-            provider_def = get_provider_by_type(
-                collections[dataset]['providers'], 'feature')
-        except ProviderTypeError:
-            try:
-                provider_def = get_provider_by_type(
-                    collections[dataset]['providers'], 'record')
-            except ProviderTypeError:
-                msg = 'Invalid provider type'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'NoApplicableCode', msg)
-
-        try:
-            p = load_plugin('provider', provider_def)
-        except ProviderGenericError as err:
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        LOGGER.debug('processing property parameters')
-        for k, v in request.params.items():
-            if k not in reserved_fieldnames and k not in p.fields.keys():
-                msg = f'unknown query parameter: {k}'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-            elif k not in reserved_fieldnames and k in p.fields.keys():
-                LOGGER.debug(f'Add property filter {k}={v}')
-                properties.append((k, v))
-
-        LOGGER.debug('processing sort parameter')
-        val = request.params.get('sortby')
-
-        if val is not None:
-            sortby = []
-            sorts = val.split(',')
-            for s in sorts:
-                prop = s
-                order = '+'
-                if s[0] in ['+', '-']:
-                    order = s[0]
-                    prop = s[1:]
-
-                if prop not in p.fields.keys():
-                    msg = 'bad sort property'
-                    return self.get_exception(
-                        HTTPStatus.BAD_REQUEST, headers, request.format,
-                        'InvalidParameterValue', msg)
-
-                sortby.append({'property': prop, 'order': order})
-        else:
-            sortby = []
-
-        LOGGER.debug('processing properties parameter')
-        val = request.params.get('properties')
-
-        if val is not None:
-            select_properties = val.split(',')
-            properties_to_check = set(p.properties) | set(p.fields.keys())
-
-            if (len(list(set(select_properties) -
-                         set(properties_to_check))) > 0):
-                msg = 'unknown properties specified'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        else:
-            select_properties = []
-
-        LOGGER.debug('processing skipGeometry parameter')
-        val = request.params.get('skipGeometry')
-        if val is not None:
-            skip_geometry = str2bool(val)
-        else:
-            skip_geometry = False
-
-        LOGGER.debug('Processing filter-crs parameter')
-        filter_crs = request.params.get('filter-crs', DEFAULT_CRS)
-        LOGGER.debug('Processing filter-lang parameter')
-        filter_lang = request.params.get('filter-lang')
-        if filter_lang != 'cql-json':  # @TODO add check from the configuration
-            msg = 'Invalid filter language'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        LOGGER.debug('Querying provider')
-        LOGGER.debug(f'offset: {offset}')
-        LOGGER.debug(f'limit: {limit}')
-        LOGGER.debug(f'resulttype: {resulttype}')
-        LOGGER.debug(f'sortby: {sortby}')
-        LOGGER.debug(f'bbox: {bbox}')
-        LOGGER.debug(f'datetime: {datetime_}')
-        LOGGER.debug(f'properties: {select_properties}')
-        LOGGER.debug(f'skipGeometry: {skip_geometry}')
-        LOGGER.debug(f'q: {q}')
-        LOGGER.debug(f'filter-lang: {filter_lang}')
-        LOGGER.debug(f'filter-crs: {filter_crs}')
-
-        LOGGER.debug('Processing headers')
-
-        LOGGER.debug('Processing request content-type header')
-        if (request_headers.get(
-            'Content-Type') or request_headers.get(
-                'content-type')) != 'application/query-cql-json':
-            msg = ('Invalid body content-type')
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidHeaderValue', msg)
-
-        LOGGER.debug('Processing body')
-
-        if not request.data:
-            msg = 'missing request data'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'MissingParameterValue', msg)
-
-        filter_ = None
-        try:
-            # Parse bytes data, if applicable
-            data = request.data.decode()
-            LOGGER.debug(data)
-        except UnicodeDecodeError as err:
-            LOGGER.error(err)
-            msg = 'Unicode error in data'
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        # FIXME: remove testing backend in use once CQL support is normalized
-        if p.name == 'PostgreSQL':
-            LOGGER.debug('processing PostgreSQL CQL_JSON data')
-            try:
-                filter_ = parse_cql_json(data)
-                filter_ = modify_pygeofilter(
-                    filter_,
-                    filter_crs_uri=filter_crs,
-                    storage_crs_uri=provider_def.get('storage_crs'),
-                    geometry_column_name=provider_def.get('geom_field')
-                )
-            except Exception as err:
-                LOGGER.error(err)
-                msg = f'Bad CQL string : {data}'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        else:
-            LOGGER.debug('processing Elasticsearch CQL_JSON data')
-            try:
-                filter_ = CQLModel.model_validate_json(data)
-            except Exception as err:
-                LOGGER.error(err)
-                msg = f'Bad CQL string : {data}'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-
-        try:
-            content = p.query(offset=offset, limit=limit,
-                              resulttype=resulttype, bbox=bbox,
-                              datetime_=datetime_, properties=properties,
-                              sortby=sortby,
-                              select_properties=select_properties,
-                              skip_geometry=skip_geometry,
-                              q=q,
-                              filterq=filter_)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        return headers, HTTPStatus.OK, to_json(content, self.pretty_print)
-
-    @gzip
-    @pre_process
-    def manage_collection_item(
-            self, request: Union[APIRequest, Any],
-            action, dataset, identifier=None) -> Tuple[dict, int, str]:
-        """
-        Adds an item to a collection
-
-        :param request: A request object
-        :param action: an action among 'create', 'update', 'delete', 'options'
-        :param dataset: dataset name
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid(PLUGINS['formatter'].keys()):
-            return self.get_format_exception(request)
-
-        # Set Content-Language to system locale until provider locale
-        # has been determined
-        headers = request.get_response_headers(SYSTEM_LOCALE,
-                                               **self.api_headers)
-
-        collections = filter_dict_by_key_value(self.config['resources'],
-                                               'type', 'collection')
-
-        if dataset not in collections.keys():
-            msg = 'Collection not found'
-            LOGGER.error(msg)
-            return self.get_exception(
-                HTTPStatus.NOT_FOUND, headers, request.format, 'NotFound', msg)
-
-        LOGGER.debug('Loading provider')
-        try:
-            provider_def = get_provider_by_type(
-                collections[dataset]['providers'], 'feature')
-            p = load_plugin('provider', provider_def)
-        except ProviderTypeError:
-            try:
-                provider_def = get_provider_by_type(
-                    collections[dataset]['providers'], 'record')
-                p = load_plugin('provider', provider_def)
-            except ProviderTypeError:
-                msg = 'Invalid provider type'
-                LOGGER.error(msg)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-
-        if action == 'options':
-            headers['Allow'] = 'HEAD, GET'
-            if p.editable:
-                if identifier is None:
-                    headers['Allow'] += ', POST'
-                else:
-                    headers['Allow'] += ', PUT, DELETE'
-            return headers, HTTPStatus.OK, ''
-
-        if not p.editable:
-            msg = 'Collection is not editable'
-            LOGGER.error(msg)
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        if action in ['create', 'update'] and not request.data:
-            msg = 'No data found'
-            LOGGER.error(msg)
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, request.format,
-                'InvalidParameterValue', msg)
-
-        if action == 'create':
-            LOGGER.debug('Creating item')
-            try:
-                identifier = p.create(request.data)
-            except TypeError as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-            except ProviderGenericError as err:
-                LOGGER.error(err)
-                return self.get_exception(
-                    err.http_status_code, headers, request.format,
-                    err.ogc_exception_code, err.message)
-
-            headers['Location'] = f'{self.get_collections_url()}/{dataset}/items/{identifier}'  # noqa
-
-            return headers, HTTPStatus.CREATED, ''
-
-        if action == 'update':
-            LOGGER.debug('Updating item')
-            try:
-                _ = p.update(identifier, request.data)
-            except TypeError as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-            except ProviderGenericError as err:
-                LOGGER.error(err)
-                return self.get_exception(
-                    err.http_status_code, headers, request.format,
-                    err.ogc_exception_code, err.message)
-
-            return headers, HTTPStatus.NO_CONTENT, ''
-
-        if action == 'delete':
-            LOGGER.debug('Deleting item')
-            try:
-                _ = p.delete(identifier)
-            except ProviderGenericError as err:
-                LOGGER.error(err)
-                return self.get_exception(
-                    err.http_status_code, headers, request.format,
-                    err.ogc_exception_code, err.message)
-
-            return headers, HTTPStatus.OK, ''
-
-    @gzip
-    @pre_process
-    def get_collection_item(self, request: Union[APIRequest, Any],
-                            dataset, identifier) -> Tuple[dict, int, str]:
-        """
-        Get a single collection item
-
-        :param request: A request object
-        :param dataset: dataset name
-        :param identifier: item identifier
-
-        :returns: tuple of headers, status code, content
-        """
-
-        if not request.is_valid():
-            return self.get_format_exception(request)
-
-        # Set Content-Language to system locale until provider locale
-        # has been determined
-        headers = request.get_response_headers(SYSTEM_LOCALE,
-                                               **self.api_headers)
-
-        LOGGER.debug('Processing query parameters')
-
-        collections = filter_dict_by_key_value(self.config['resources'],
-                                               'type', 'collection')
-
-        if dataset not in collections.keys():
-            msg = 'Collection not found'
-            return self.get_exception(
-                HTTPStatus.NOT_FOUND, headers, request.format, 'NotFound', msg)
-
-        LOGGER.debug('Loading provider')
-
-        try:
-            provider_type = 'feature'
-            provider_def = get_provider_by_type(
-                collections[dataset]['providers'], provider_type)
-            p = load_plugin('provider', provider_def)
-        except ProviderTypeError:
-            try:
-                provider_type = 'record'
-                provider_def = get_provider_by_type(
-                    collections[dataset]['providers'], provider_type)
-                p = load_plugin('provider', provider_def)
-            except ProviderTypeError:
-                msg = 'Invalid provider type'
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        crs_transform_spec = None
-        if provider_type == 'feature':
-            # crs query parameter is only available for OGC API - Features
-            # right now, not for OGC API - Records.
-            LOGGER.debug('Processing crs parameter')
-            query_crs_uri = request.params.get('crs')
-            try:
-                crs_transform_spec = self._create_crs_transform_spec(
-                    provider_def, query_crs_uri,
-                )
-            except (ValueError, CRSError) as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, request.format,
-                    'InvalidParameterValue', msg)
-            self._set_content_crs_header(headers, provider_def, query_crs_uri)
-
-        # Get provider language (if any)
-        prv_locale = l10n.get_plugin_locale(provider_def, request.raw_locale)
-
-        try:
-            LOGGER.debug(f'Fetching id {identifier}')
-            content = p.get(
-                identifier,
-                language=prv_locale,
-                crs_transform_spec=crs_transform_spec,
-            )
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        if content is None:
-            msg = 'identifier not found'
-            return self.get_exception(HTTPStatus.BAD_REQUEST, headers,
-                                      request.format, 'NotFound', msg)
-
-        uri = content['properties'].get(p.uri_field) if p.uri_field else \
-            f'{self.get_collections_url()}/{dataset}/items/{identifier}'
-
-        if 'links' not in content:
-            content['links'] = []
-
-        content['links'].extend([{
-            'type': FORMAT_TYPES[F_JSON],
-            'rel': 'root',
-            'title': 'The landing page of this server as JSON',
-            'href': f"{self.base_url}?f={F_JSON}"
-            }, {
-            'type': FORMAT_TYPES[F_HTML],
-            'rel': 'root',
-            'title': 'The landing page of this server as HTML',
-            'href': f"{self.base_url}?f={F_HTML}"
-            }, {
-            'rel': request.get_linkrel(F_JSON),
-            'type': 'application/geo+json',
-            'title': 'This document as GeoJSON',
-            'href': f'{uri}?f={F_JSON}'
-            }, {
-            'rel': request.get_linkrel(F_JSONLD),
-            'type': FORMAT_TYPES[F_JSONLD],
-            'title': 'This document as RDF (JSON-LD)',
-            'href': f'{uri}?f={F_JSONLD}'
-            }, {
-            'rel': request.get_linkrel(F_HTML),
-            'type': FORMAT_TYPES[F_HTML],
-            'title': 'This document as HTML',
-            'href': f'{uri}?f={F_HTML}'
-            }, {
-            'rel': 'collection',
-            'type': FORMAT_TYPES[F_JSON],
-            'title': l10n.translate(collections[dataset]['title'],
-                                    request.locale),
-            'href': f'{self.get_collections_url()}/{dataset}'
-        }])
-
-        link_request_format = (
-            request.format if request.format is not None else F_JSON
-        )
-        if 'prev' in content:
-            content['links'].append({
-                'rel': 'prev',
-                'type': FORMAT_TYPES[link_request_format],
-                'href': f"{self.get_collections_url()}/{dataset}/items/{content['prev']}?f={link_request_format}"  # noqa
-            })
-        if 'next' in content:
-            content['links'].append({
-                'rel': 'next',
-                'type': FORMAT_TYPES[link_request_format],
-                'href': f"{self.get_collections_url()}/{dataset}/items/{content['next']}?f={link_request_format}"  # noqa
-            })
-
-        # Set response language to requested provider locale
-        # (if it supports language) and/or otherwise the requested pygeoapi
-        # locale (or fallback default locale)
-        l10n.set_response_language(headers, prv_locale, request.locale)
-
-        if request.format == F_HTML:  # render
-            content['title'] = l10n.translate(collections[dataset]['title'],
-                                              request.locale)
-            content['id_field'] = p.id_field
-            if p.uri_field is not None:
-                content['uri_field'] = p.uri_field
-            if p.title_field is not None:
-                content['title_field'] = l10n.translate(p.title_field,
-                                                        request.locale)
-            content['collections_path'] = self.get_collections_url()
-
-            content = render_j2_template(self.tpl_config,
-                                         'collections/items/item.html',
-                                         content, request.locale)
-            return headers, HTTPStatus.OK, content
-
-        elif request.format == F_JSONLD:
-            content = geojson2jsonld(
-                self, content, dataset, uri, (p.uri_field or 'id')
-            )
-
-        return headers, HTTPStatus.OK, to_json(content, self.pretty_print)
-
-    @pre_process
-    @jsonldify
-    def get_collection_coverage(self, request: Union[APIRequest, Any],
-                                dataset) -> Tuple[dict, int, str]:
-        """
-        Returns a subset of a collection coverage
-
-        :param request: A request object
-        :param dataset: dataset name
-
-        :returns: tuple of headers, status code, content
-        """
-
-        query_args = {}
-        format_ = request.format or F_JSON
-
-        # Force response content type and language (en-US only) headers
-        headers = request.get_response_headers(SYSTEM_LOCALE,
-                                               **self.api_headers)
-
-        LOGGER.debug('Loading provider')
-        try:
-            collection_def = get_provider_by_type(
-                self.config['resources'][dataset]['providers'], 'coverage')
-
-            p = load_plugin('provider', collection_def)
-        except KeyError:
-            msg = 'collection does not exist'
-            return self.get_exception(
-                HTTPStatus.NOT_FOUND, headers, format_,
-                'InvalidParameterValue', msg)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        LOGGER.debug('Processing bbox parameter')
-
-        bbox = request.params.get('bbox')
-
-        if bbox is None:
-            bbox = []
-        else:
-            try:
-                bbox = validate_bbox(bbox)
-            except ValueError as err:
-                msg = str(err)
-                return self.get_exception(
-                    HTTPStatus.INTERNAL_SERVER_ERROR, headers, format_,
-                    'InvalidParameterValue', msg)
-
-        query_args['bbox'] = bbox
-
-        LOGGER.debug('Processing bbox-crs parameter')
-
-        bbox_crs = request.params.get('bbox-crs')
-        if bbox_crs is not None:
-            query_args['bbox_crs'] = bbox_crs
-
-        LOGGER.debug('Processing datetime parameter')
-
-        datetime_ = request.params.get('datetime')
-
-        try:
-            datetime_ = validate_datetime(
-                self.config['resources'][dataset]['extents'], datetime_)
-        except ValueError as err:
-            msg = str(err)
-            return self.get_exception(
-                HTTPStatus.BAD_REQUEST, headers, format_,
-                'InvalidParameterValue', msg)
-
-        query_args['datetime_'] = datetime_
-        query_args['format_'] = format_
-
-        properties = request.params.get('properties')
-        if properties:
-            LOGGER.debug('Processing properties parameter')
-            query_args['properties'] = [rs for
-                                        rs in properties.split(',') if rs]
-            LOGGER.debug(f"Fields: {query_args['properties']}")
-
-            for a in query_args['properties']:
-                if a not in p.fields:
-                    msg = 'Invalid field specified'
-                    return self.get_exception(
-                        HTTPStatus.BAD_REQUEST, headers, format_,
-                        'InvalidParameterValue', msg)
-
-        if 'subset' in request.params:
-            LOGGER.debug('Processing subset parameter')
-            try:
-                subsets = validate_subset(request.params['subset'] or '')
-            except (AttributeError, ValueError) as err:
-                msg = f'Invalid subset: {err}'
-                LOGGER.error(msg)
-                return self.get_exception(
-                        HTTPStatus.BAD_REQUEST, headers, format_,
-                        'InvalidParameterValue', msg)
-
-            if not set(subsets.keys()).issubset(p.axes):
-                msg = 'Invalid axis name'
-                LOGGER.error(msg)
-                return self.get_exception(
-                    HTTPStatus.BAD_REQUEST, headers, format_,
-                    'InvalidParameterValue', msg)
-
-            query_args['subsets'] = subsets
-            LOGGER.debug(f"Subsets: {query_args['subsets']}")
-
-        LOGGER.debug('Querying coverage')
-        try:
-            data = p.query(**query_args)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        mt = collection_def['format']['name']
-        if format_ == mt:  # native format
-            if p.filename is not None:
-                cd = f'attachment; filename="{p.filename}"'
-                headers['Content-Disposition'] = cd
-
-            headers['Content-Type'] = collection_def['format']['mimetype']
-            return headers, HTTPStatus.OK, data
-        elif format_ == F_JSON:
-            headers['Content-Type'] = 'application/prs.coverage+json'
-            return headers, HTTPStatus.OK, to_json(data, self.pretty_print)
-        else:
-            return self.get_format_exception(request)
-
-    @gzip
-    def get_collection_map_legend(
-            self, request: Union[APIRequest, Any],
-            dataset, style=None) -> Tuple[dict, int, str]:
-        """
-        Returns a subset of a collection map legend
-
-        :param request: A request object
-        :param dataset: dataset name
-        :param style: style name
-
-        :returns: tuple of headers, status code, content
-        """
-
-        format_ = 'png'
-        headers = request.get_response_headers(**self.api_headers)
-        LOGGER.debug('Processing query parameters')
-
-        LOGGER.debug('Loading provider')
-        try:
-            collection_def = get_provider_by_type(
-                self.config['resources'][dataset]['providers'], 'map')
-
-            p = load_plugin('provider', collection_def)
-        except KeyError:
-            exception = {
-                'code': 'InvalidParameterValue',
-                'description': 'collection does not exist'
-            }
-            LOGGER.error(exception)
-            return headers, HTTPStatus.NOT_FOUND, to_json(
-                exception, self.pretty_print)
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        LOGGER.debug('Generating legend')
-        try:
-            data = p.get_legend(style, request.params.get('f', 'png'))
-        except ProviderGenericError as err:
-            LOGGER.error(err)
-            return self.get_exception(
-                err.http_status_code, headers, request.format,
-                err.ogc_exception_code, err.message)
-
-        mt = collection_def['format']['name']
-
-        if format_ == mt:
-            headers['Content-Type'] = collection_def['format']['mimetype']
-            return headers, HTTPStatus.OK, data
-        else:
-            exception = {
-                'code': 'InvalidParameterValue',
-                'description': 'invalid format parameter'
-            }
-            LOGGER.error(exception)
-            return headers, HTTPStatus.BAD_REQUEST, to_json(
-                data, self.pretty_print)
 
     def get_exception(self, status, headers, format_, code,
                       description) -> Tuple[dict, int, str]:
